blackjack: represent game state
Created a Hand struct to represent a hand of cards, a Player struct to represent a player's status in the game, including cards, score and if he has bust, and a Game struct representing the entire game including the participating players and order of play.
Starting a new game creates an empty (for now) hand for every player.