blackjack: hide hands not belonging to current player
To keep the game interesting, we do not allow players to see other players' cards, with the exception that everyone can see of the dealer's cards.

We use a field on the request for the player name for simplicity, we should be using proper authentication instead to get the actual player name.